requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pygame>=2.0.0
torch>=1.7.0
pyyaml>=5.4.0
matplotlib>=3.3.0
